workspace/workspace-01/pyproject.toml
[project]
name = "workspace-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "strawberry-graphql>=0.312.2",
    "uvicorn>=0.42.0",
]
